insurance_agent_api/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
langchain-core>=0.3.36,<2.0.0
langchain-google-genai>=2.0.0
langgraph>=0.2.0
pydantic>=2.0.0
python-dotenv>=1.0.0
